Add AutomationService and make intervals double
Change Automation.Interval and Elapsed from int to double.
Add AutomationService, purchase logic ,update loop.
ExecuteAutomation when interval reached

+++ Keyboard clicker/Models/Automation.cs
@@ -13,8 +13,8 @@ namespace Keyboard_clicker.Models
         public double Price { get; set; }
         public int Quantity { get; set; }
 
-        public int Interval { get; set; }
-        public int Elapsed { get; set; }
+        public double Interval { get; set; }
+        public double Elapsed { get; set; }
 
         public string AutomationType { get; set; }
     }
